[sources/_template.py] -> update comments
